+++ src/jquery.restrictedtextfield.js
@@ -94,33 +94,33 @@
 
          var dest = $.fn.restrictedTextField.types = $.fn.restrictedTextField.types || [];
 
-         _addType( dest, "alpha",                   /^[a-zA-Z]*$/            , null );
-         _addType( dest, "upperAlpha",              /^[A-Z]*$/               , null );
-         _addType( dest, "lowerAlpha",              /^[a-z]*$/               , null );
-         _addType( dest, "alphaSpace",              /^[a-zA-Z\ ]*$/          , null );
-         _addType( dest, "upperAlphaSpace",         /^[A-Z\ ]*$/             , null );
-         _addType( dest, "lowerAlphaSpace",         /^[a-z\ ]*$/             , null );
-         _addType( dest, "alphanumeric",            /^[a-zA-Z\d]*$/          , null );
-         _addType( dest, "upperAlphanumeric",       /^[A-Z\d]*$/             , null );
-         _addType( dest, "lowerAlphanumeric",       /^[a-z\d]*$/             , null );
-         _addType( dest, "alphanumericSpace",       /^[a-zA-Z\d\ ]*$/        , null );
-         _addType( dest, "upperAlphanumericSpace",  /^[A-Z\d\ ]*$/           , null );
-         _addType( dest, "lowerAlphanumericSpace",  /^[a-z\d\ ]*$/           , null );
-         _addType( dest, "int",                     /^-?\d*$/                , partialNegativeInt );
-         _addType( dest, "positiveInt",             /^\d*$/                  , null );
-         _addType( dest, "negativeInt",             /^0*$|^-\d*$/            , partialNegativeInt );
-         _addType( dest, "strictInt",               /^0*$|^-?0*[1-9]\d*$/    , partialNegativeInt );
-         _addType( dest, "strictPositiveInt",       /^0*$|^0*[1-9]\d*$/      , null );
-         _addType( dest, "strictNegativeInt",       /^0*$|^-0*[1-9]\d*$/     , partialNegativeInt );
+         _addType( dest, "alpha",                   /^[a-zA-Z]+$/            , null );
+         _addType( dest, "upperAlpha",              /^[A-Z]+$/               , null );
+         _addType( dest, "lowerAlpha",              /^[a-z]+$/               , null );
+         _addType( dest, "alphaSpace",              /^[a-zA-Z\ ]+$/          , null );
+         _addType( dest, "upperAlphaSpace",         /^[A-Z\ ]+$/             , null );
+         _addType( dest, "lowerAlphaSpace",         /^[a-z\ ]+$/             , null );
+         _addType( dest, "alphanumeric",            /^[a-zA-Z\d]+$/          , null );
+         _addType( dest, "upperAlphanumeric",       /^[A-Z\d]+$/             , null );
+         _addType( dest, "lowerAlphanumeric",       /^[a-z\d]+$/             , null );
+         _addType( dest, "alphanumericSpace",       /^[a-zA-Z\d\ ]+$/        , null );
+         _addType( dest, "upperAlphanumericSpace",  /^[A-Z\d\ ]+$/           , null );
+         _addType( dest, "lowerAlphanumericSpace",  /^[a-z\d\ ]+$/           , null );
+         _addType( dest, "int",                     /^-?\d+$/                , partialNegativeInt );
+         _addType( dest, "positiveInt",             /^\d+$/                  , null );
+         _addType( dest, "negativeInt",             /^0+$|^-\d+$/            , partialNegativeInt );
+         _addType( dest, "strictInt",               /^0+$|^-?0*[1-9]\d*$/    , partialNegativeInt );
+         _addType( dest, "strictPositiveInt",       /^\d+$/                  , null );
+         _addType( dest, "strictNegativeInt",       /^0+$|^-0*[1-9]\d*$/     , partialNegativeInt );
          _addType( dest, "float",                   /^-?\d*\.?\d+$/          , partialPosNegFloat );
          _addType( dest, "positiveFloat",           /^\d*\.?\d+$/            , partialPositiveFloat );
          _addType( dest, "negativeFloat",           /^-?0+$|^-?0*\.?0+$|^-\d*\.?\d+$/ , partialNegativeFloat );
          _addType( dest, "strictFloat",             /^0*\.0+$|^-?0*\.\d*[1-9]$|^-?\d*\.\d*[1-9]\d*$|^-?0*[1-9]\d*?\d*\.\d+$/  , partialPosNegFloat );
          _addType( dest, "strictPositiveFloat",     /^0*\.0+$|^0*\.\d*[1-9]$|^\d*\.\d*[1-9]\d*$|^\d*[1-9]*\d*\.\d+$/     , partialPositiveFloat );
          _addType( dest, "strictNegativeFloat",     /^0*\.0+$|^-0*\.\d*[1-9]$|^-\d*\.\d*[1-9]\d*$|^-[^0]\d*?\d*\.\d+$/   , partialNegativeFloat );
-         _addType( dest, "money",                   /^-?\d*\.?\d{2}$/      , partialPosNegMoney );
-         _addType( dest, "positiveMoney",           /^\d*\.?\d{2}$/        , partialPositiveMoney );
-         _addType( dest, "negativeMoney",           /^-\d*\.?\d{2}$/       , partialNegativeMoney );
+         _addType( dest, "money",                   /^-?\d*\.?\d{2}$/        , partialPosNegMoney );
+         _addType( dest, "positiveMoney",           /^\d*\.?\d{2}$/          , partialPositiveMoney );
+         _addType( dest, "negativeMoney",           /^-\d*\.?\d{2}$/         , partialNegativeMoney );
 
          // Positive floating-point numbers with one or two numbers after the decimal point;
          // Positive integers;
@@ -128,7 +128,7 @@
          // wrapping the value in parentheses rather than using a minus sign (sometimes referred to as "accounting notation")
          _addType( dest, "accountingMoney", /^\d*\.?\d{1,2}$|^\(\d*\.?\d{1,2}\)$/  ,  /^[\.\d]$|^\.$|^\d*\.$|^\(\d*\.?$|^\(\d*\.\d{1,2}?$/ );
 
-         _addType( dest, "negativeAccountingMoney", /^\(\d*\.?\d{1,2}\)$/  ,  /^\(\d*\.?$|^\(\d*\.\d{1,2}?$/ );  // The negative-only version of accountingMoney
+         _addType( dest, "negativeAccountingMoney", /^\(\d*\.?\d{2}\)$/  ,  /^\(\d*\.?$|^\(\d*\.\d{1,2}?$/ );  // The negative-only version of accountingMoney
       }
 
       var regexes = $.fn.restrictedTextField.types[ settings.type ];
@@ -202,9 +202,12 @@
          } );
 
          jqThis.on( "blur", function() {
-            applyFormatting( this );
-
             var responseEvent = EVENT_VALIDATION_FAILED;
+
+            if( settings.type === "money" || settings.type === "positiveMoney" || settings.type === "negativeMoney" ||
+                settings.type === "accountingMoney" || settings.type ==="negativeAccountingMoney" ) {
+               formatMoney( this );
+            }
 
             if( this.value.length === 0 || regexes.fullRegex.test(this.value) ) {
                responseEvent = EVENT_VALIDATION_SUCCESS;
@@ -215,44 +218,48 @@
             jqThis.trigger( responseEvent );
          } );
 
-         function applyFormatting( domField ) {
+         function formatMoney( domField ) {
             if( isNothing(domField.value) ) return;
 
-            var t = settings.type;
+            if( (/^-?\d*\.\d\d?$/).test(domField.value) ||
+                (/^\(\d*\.?\d{1,2}\)$/).test(domField.value) ||
+                $.fn.restrictedTextField.types["int"].fullRegex.test(domField.value) ) {
 
-            if( t === "money" || t === "positiveMoney" || t === "negativeMoney" || t === "accountingMoney" || t ==="negativeAccountingMoney" ) {
+               var formatted = domField.value;
+
+               domField.value = trimLeadingZero( domField.value );
+
                var len         = domField.value.length;
+               var sign        = domField.value[0] === "-" ? "-" : "";
                var openParen   = domField.value[0] === "(" ? "(" : "";
                var closeParen  = domField.value[ len - 1 ] === ")" ? ")" : ""; 
                var decimalIdx  = domField.value.indexOf( "." );
-               var integerPart = domField.value.substring( openParen === "" ? 0 : 1,
+               var integerPart = domField.value.substring( openParen === "" && sign === "" ? 0 : 1,
                                                            decimalIdx > 0 ? decimalIdx -1 : (closeParen === ")" ? len - 1 : len) );
                var decimalPart = "00";
 
                if( decimalIdx > -1 ) {
-                  integerPart = domField.value.substring( openParen === "" ? 0 : 1, decimalIdx );
+                  integerPart = domField.value.substring( openParen === "" && sign === "" ? 0 : 1, decimalIdx );
                   decimalPart = domField.value.substring( decimalIdx + 1, closeParen === "" ? len : len - 1 );
                }
 
-               integerPart = trimLeadingZero( integerPart );
-               domField.value = openParen + integerPart + "." + decimalPart + (decimalPart.length === 1 ? "0" : "") + closeParen;
+               if( (integerPart === "" || parseInt(integerPart) === 0) && (parseInt(decimalPart) === 0) ) {
+                  formatted = "0.00"
+               } else {
+                  formatted = openParen
+                            + sign
+                            + (integerPart === "" ? "0" : integerPart)
+                            + "."
+                            + decimalPart
+                            + (decimalPart.length === 1 ? "0" : "")
+                            + closeParen;
+               }
+
+               domField.value = formatted;
             }
 
             function trimLeadingZero( str ) {
-               var trim = "";
-               var startAt = -1;
-
-               for( var i = 0; i < str.length; i++ ) {
-                  if( str[i] !== "0" ) {
-                     break;
-                  }
-
-                  startAt = i;
-               }
-
-               trim = startAt >= 0 ? str.substring( startAt + 1 ) : str;
-
-               return trim;
+               return str.replace( /^(-?)0*(\d+\.\d+)/, "$1$2" );
             }
          }
       } );